# web/openapi.json
{
  "components": {
    "schemas": {
...
      "AssetIngestRequest": {
        "additionalProperties": false,
        "description": "扫盘 / 入库请求。\n\n``kind`` 留空 = 三类都扫（面板「扫全部」按钮）。``ids`` 留空 = 该目录全扫，\n给了 = 只处理这几条（「重扫这一条」按钮，坏文件修好之后不用整批重来）。\n\n``license`` 只对**本次新入库**的条目生效；已入库的按行里存的那份走 ——\n否则一次「顺手全扫」会把每条素材的授权都改成本次请求里填的那个。",
        "properties": {
          "dry_run": {
            "default": false,
            "title": "Dry Run",
            "type": "boolean"
          },
          "ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ids"
          },
          "kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "AssetIngestRequest",
        "type": "object"
      },
      "AssetKind": {
        "description": "三类素材（与 §3.3.14 的两张表 + §4.3.1 的音色目录一一对应）。",
        "enum": [
          "broll",
          "voice",
          "bgm"
        ],
        "title": "AssetKind",
        "type": "string"
      },
...
      "AssetPatchRequest": {
        "additionalProperties": false,
        "description": "改一条素材（启用 / 停用 / 授权 / 标签 / 可用区间 / 情绪…）。\n\n``kind`` 可以不给：服务层按 id 跨类反查（同一个 id 在两类里都有 ⇒ 明确报错，\n绝不猜一个然后改错行）。\n\n字段级约束（``license`` 的枚举、可用区间非空、``usable_from_ms ≥ 0``）**不在这里**：\n它们住在服务层 :func:`~studio.services.asset_service._clean_fields`。请求体再抄一份，\n就是「面板放行、保存却 422」的来源（裁定：真判定只有一次）。",
        "properties": {
          "bpm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bpm"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "has_text": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has Text"
          },
          "kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "licensed_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Licensed To"
          },
          "loopable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loopable"
          },
          "mood": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mood"
          },
          "proof_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proof Path"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "text_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Path"
          },
          "usable_from_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usable From Ms"
          },
          "usable_to_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usable To Ms"
          }
        },
        "title": "AssetPatchRequest",
        "type": "object"
      },
...
      "BgmItemModel": {
        "additionalProperties": false,
        "description": "一条 BGM。",
        "properties": {
          "bitrate_kbps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bitrate Kbps"
          },
          "bpm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bpm"
          },
          "channels": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channels"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "const": "bgm",
            "title": "Kind",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "license": {
            "title": "License",
            "type": "string"
          },
          "licensed_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Licensed To"
          },
          "loopable": {
            "title": "Loopable",
            "type": "boolean"
          },
          "loudness_lufs": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loudness Lufs"
          },
          "mood": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mood"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "proof_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proof Path"
          },
          "sample_rate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Rate"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "use_count": {
            "title": "Use Count",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "id",
          "path",
          "duration_ms",
          "sample_rate",
          "channels",
          "bitrate_kbps",
          "loudness_lufs",
          "bpm",
          "mood",
          "tags",
          "loopable",
          "license",
          "source_url",
          "proof_path",
          "licensed_to",
          "use_count",
          "last_used_at",
          "enabled",
          "created_at"
        ],
        "title": "BgmItemModel",
        "type": "object"
      },
...
      "BrollItemModel": {
        "additionalProperties": false,
        "description": "一条跑酷素材（``kind`` 是判别字段，下同）。",
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "fps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fps"
          },
          "has_text": {
            "title": "Has Text",
            "type": "boolean"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "const": "broll",
            "title": "Kind",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "license": {
            "title": "License",
            "type": "string"
          },
          "licensed_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Licensed To"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "proof_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proof Path"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "thumb_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumb Path"
          },
          "usable_from_ms": {
            "title": "Usable From Ms",
            "type": "integer"
          },
          "usable_to_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usable To Ms"
          },
          "use_count": {
            "title": "Use Count",
            "type": "integer"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "required": [
          "kind",
          "id",
          "path",
          "duration_ms",
          "usable_from_ms",
          "usable_to_ms",
          "width",
          "height",
          "fps",
          "has_text",
          "tags",
          "license",
          "thumb_path",
          "source_url",
          "proof_path",
          "licensed_to",
          "use_count",
          "last_used_at",
          "enabled",
          "created_at"
        ],
        "title": "BrollItemModel",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScanReportModel": {
        "additionalProperties": false,
        "description": "一次扫盘（``dry_run=true``）或一次入库（``dry_run=false``）的完整结论。\n\n**扫盘是只读的**：``dry_run=true`` 一个字节都不写库，面板「先看看会怎样」用它。",
        "properties": {
          "dry_run": {
            "title": "Dry Run",
            "type": "boolean"
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/ScanSectionModel"
            },
            "title": "Sections",
            "type": "array"
          },
          "totals": {
            "$ref": "#/components/schemas/ScanTotalsModel"
          }
        },
        "required": [
          "dry_run",
          "license",
          "sections",
          "totals"
        ],
        "title": "ScanReportModel",
        "type": "object"
      },
      "ScanSectionModel": {
        "additionalProperties": false,
        "description": "一类素材的扫盘结果。\n\n``strays`` 是「目录里存在、但命名不合规」的文件：**如实报出来，绝不悄悄忽略**\n—— 用户把 ``跑酷1.mp4`` 丢进去，看到的应该是「它没被认出来，改名成\n``parkour_xxx.mp4``」，而不是「扫了 0 条」。",
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/ScannedAssetModel"
            },
            "title": "Assets",
            "type": "array"
          },
          "counts": {
            "$ref": "#/components/schemas/ScanCountsModel"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "missing": {
            "items": {
              "type": "string"
            },
            "title": "Missing",
            "type": "array"
          },
          "root": {
            "title": "Root",
            "type": "string"
          },
          "root_missing": {
            "title": "Root Missing",
            "type": "boolean"
          },
          "stats": {
            "$ref": "#/components/schemas/AssetStatsModel"
          },
          "strays": {
            "items": {
              "type": "string"
            },
            "title": "Strays",
            "type": "array"
          }
        },
        "required": [
          "kind",
          "root",
          "root_missing",
          "assets",
          "strays",
          "missing",
          "stats",
          "counts"
        ],
        "title": "ScanSectionModel",
        "type": "object"
      },
...
      "VoiceItemModel": {
        "additionalProperties": false,
        "description": "一个零样本音色（``path`` 是**目录**：``ref_NN`` + ``ref.txt`` + ``profile.json``）。",
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "const": "voice",
            "title": "Kind",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "licensed_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Licensed To"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "peak_db": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Db"
          },
          "proof_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proof Path"
          },
          "ref_count": {
            "title": "Ref Count",
            "type": "integer"
          },
          "sample_rate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Rate"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "text_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Path"
          },
          "total_duration_ms": {
            "title": "Total Duration Ms",
            "type": "integer"
          },
          "use_count": {
            "title": "Use Count",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "id",
          "path",
          "ref_count",
          "total_duration_ms",
          "sample_rate",
          "peak_db",
          "text_path",
          "proof_path",
          "license",
          "source_url",
          "licensed_to",
          "use_count",
          "last_used_at",
          "enabled",
          "created_at"
        ],
        "title": "VoiceItemModel",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AI 全自动营销号制片台",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/assets/ingest": {
      "post": {
        "description": "扫盘 / 入库（``dry_run=true`` ⇒ **一个字节都不写库**）。\n\n``license`` 只对本次新入库的条目生效；已入库的按行里存的那份走 —— 否则一次\n「顺手全扫」会把每条素材的授权都改成本次请求里填的那个。",
        "operationId": "ingest_assets_api_v1_assets_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanReportModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Assets",
        "tags": [
          "assets"
        ]
      }
    },
...
    "/api/v1/assets/{asset_id}": {
      "patch": {
        "description": "改一条素材（启用 / 停用 / 授权 / 标签 / 可用区间 / 情绪…），并留痕。\n\n``enabled`` 走的是**同一条** :meth:`AssetService.set_enabled` 路径（幂等：\n状态没变就不留痕），而不是这里再写一遍「赋值 + 记日志」。\n\n返回的是**行字典**，由 ``response_model``（``kind`` 判别的联合）复核 ——\n路由这一层不重复实现「三类各自长什么样」，那是契约的事。",
        "operationId": "patch_asset_api_v1_assets__asset_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[a-z0-9][a-z0-9_\\-]{0,63}$",
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "bgm": "#/components/schemas/BgmItemModel",
                      "broll": "#/components/schemas/BrollItemModel",
                      "voice": "#/components/schemas/VoiceItemModel"
                    },
                    "propertyName": "kind"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/BrollItemModel"
                    },
                    {
                      "$ref": "#/components/schemas/BgmItemModel"
                    },
                    {
                      "$ref": "#/components/schemas/VoiceItemModel"
                    }
                  ],
                  "title": "Response Patch Asset Api V1 Assets  Asset Id  Patch"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Asset",
        "tags": [
          "assets"
        ]
      }
    },
    "/api/v1/assets/{asset_id}/media": {
      "get": {
        "description": "原文件（BGM 试听 / 跑酷预览）。\n\n路径**只从库里取**（``row.path``），绝不拼请求参数 —— 素材 id 是白名单正则，\n但真正的路径来自 DB 行，这中间没有一处能让 ``../`` 生效。\n\n音色不在这里给：``voice_profiles.path`` 是**目录**（``ref_NN`` + ``ref.txt`` +\n``profile.json``），不是能播的文件。面板按条目里的路径逐段播放。",
        "operationId": "get_asset_media_api_v1_assets__asset_id__media_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[a-z0-9][a-z0-9_\\-]{0,63}$",
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Asset Media",
        "tags": [
          "assets"
        ]
      }
    },
    "/api/v1/assets/{asset_id}/thumb": {
      "get": {
        "description": "缩略图（跑酷素材的抽帧图；**没有就 404**，不临时现抽 —— 那是入库该干的事）。",
        "operationId": "get_asset_thumb_api_v1_assets__asset_id__thumb_get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[a-z0-9][a-z0-9_\\-]{0,63}$",
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Asset Thumb",
        "tags": [
          "assets"
        ]
      }
    },
...
  }
}